Compare ratios as floating-point when determining panorama orientation
These values are all integers, so we often end up comparing `1` and `1`
when trying to figure out which way should the panorama view scroll.

+++ src/main.c
@@ -1325,7 +1325,7 @@ if (!zoom_panorama)
   }
 else
   {
-    if ((width/scrnwide)>(height/scrnhigh))
+    if (((float)width/scrnwide)>((float)height/scrnhigh))
       /* pan horizontally */
       zoom_panorama_sb=0,*swp=(scrnhigh*width)/height,*shp=scrnhigh;
     else
